displaydoc

Module expand

Enums

UseGlobalPrefix ğŸ”’

Functions

add_display_constraint_to_type_predicate ğŸ”’

Add a requirement for core::fmt::Display to a where predicate for some type.

append_where_clause_type_predicate ğŸ”’

Push new_type_predicate onto the end of where_clause.

derive ğŸ”’

ensure_display_in_where_clause_for_type ğŸ”’

Hygienically add where _: Display to the set of TypeParamBounds for ident, creating such a set if necessary.

ensure_where_clause_has_display_for_all_unconstrained_members ğŸ”’

For all declared type parameters, add a core::fmt::Display constraint, unless the type parameter already has any type constraint.

extract_trait_constraints_from_source ğŸ”’

Map each declared generic type parameter to the set of all trait boundaries declared on it.

generate_where_clause ğŸ”’

Generate a where clause that ensures all generic type parameters impl core::fmt::Display unless already constrained.

impl_enum ğŸ”’

impl_struct ğŸ”’

join_paths ğŸ”’

Create a path with segments composed of [Idents] without any PathArguments.

new_empty_where_clause ğŸ”’

Create a where clause that we can add WherePredicates to.

new_empty_where_type_predicate ğŸ”’

Create a where predicate for ident, without any bounds yet.

specialization ğŸ”’